Module dalvik::bytecode

Representation of the Dalvik bytecodes and utilities to decode them

Structs

ByteCodeDecoder
Implementations of the distinct bytecodes data layouts.

Enums

ArrayOperation

BinaryOperation

ByteCode

CompareType

InvokeKind

TestType

UnaryOperation

Type Definitions

CallSiteReference
Call site index on the Dex call site table

ClassReference
Class index on the Dex class table

FieldReference
Field index on the Dex field table

MethodReference
Method index on the Dex method table

PrototypeReference
Prototype index on the Dex prototype table

StringReference
String index on the Dex string table

TypeReference
Type index on the Dex type table