Struct dalvik::header::Header

pub struct Header {
    magic: [u8; 8],
    checksum: u32,
    signature: [u8; 20],
    file_size: u32,
    header_size: u32,
    endian_tag: u32,
    link_size: u32,
    link_offset: Option<u32>,
    map_offset: u32,
    string_ids_size: u32,
    string_ids_offset: Option<u32>,
    type_ids_size: u32,
    type_ids_offset: Option<u32>,
    prototype_ids_size: u32,
    prototype_ids_offset: Option<u32>,
    field_ids_size: u32,
    field_ids_offset: Option<u32>,
    method_ids_size: u32,
    method_ids_offset: Option<u32>,
    class_defs_size: u32,
    class_defs_offset: Option<u32>,
    data_size: u32,
    data_offset: u32,
}

Dex header representantion structure.

Fields
magic: [u8; 8] checksum: u32 signature: [u8; 20] file_size: u32 header_size: u32 endian_tag: u32 link_size: u32 link_offset: Option<u32> map_offset: u32 string_ids_size: u32 string_ids_offset: Option<u32> type_ids_size: u32 type_ids_offset: Option<u32> prototype_ids_size: u32 prototype_ids_offset: Option<u32> field_ids_size: u32 field_ids_offset: Option<u32> method_ids_size: u32 method_ids_offset: Option<u32> class_defs_size: u32 class_defs_offset: Option<u32> data_size: u32 data_offset: u32
Methods

impl Header

fn from_file<P: AsRef<Path>>(path: P) -> Result<Header>

Obtains the header from a Dex file.

fn from_reader<R: Read>(reader: R) -> Result<Header>

Obtains the header from a Dex file reader.

fn read_data<R: Read, E: ByteOrder>(reader: R,
magic: [u8; 8],
checksum: u32,
signature: [u8; 20],
file_size: u32,
header_size: u32,
endian_tag: u32)
-> Result<Header>

fn is_magic_valid(magic: &[u8; 8]) -> bool

Checks if the dex magic number given is valid.

fn get_magic(&self) -> &[u8; 8]

Gets the magic value.

fn get_dex_version(&self) -> u8

Gets Dex version.

fn get_checksum(&self) -> u32

Gets file checksum.

fn get_signature(&self) -> &[u8; 20]

Gets file SHA-1 signature.

fn get_file_size(&self) -> u32

Gets file size.

fn get_header_size(&self) -> u32

Gets header size, in bytes.

This must be 0x70.

fn get_endian_tag(&self) -> u32

Gets the endian tag.

This must be ENDIAN_CONSTANT or REVERSE_ENDIAN_CONSTANT.

fn is_little_endian(&self) -> bool

Gets wether the file is in little endian or not.

fn is_big_endian(&self) -> bool

Gets wether the file is in big endian or not.

fn get_link_size(&self) -> u32

Gets the link section size

fn get_link_offset(&self) -> Option<u32>

Gets the link section offset.

fn get_map_offset(&self) -> u32

Gets the map section offset.

fn get_string_ids_size(&self) -> u32

Gets the string IDs list size.

fn get_string_ids_offset(&self) -> Option<u32>

Gets the string IDs list offset.

fn get_type_ids_size(&self) -> u32

Gets the type IDs list size.

fn get_type_ids_offset(&self) -> Option<u32>

Gets the type IDs list offset.

fn get_prototype_ids_size(&self) -> u32

Gets the prototype IDs list size.

fn get_prototype_ids_offset(&self) -> Option<u32>

Gets the prototype IDs list offset.

fn get_field_ids_size(&self) -> u32

Gets the field IDs list size.

fn get_field_ids_offset(&self) -> Option<u32>

Gets the field IDs list offset.

fn get_method_ids_size(&self) -> u32

Gets the method IDs list size.

fn get_method_ids_offset(&self) -> Option<u32>

Gets the method IDs list offset.

fn get_class_defs_size(&self) -> u32

Gets the class definition list size.

fn get_class_defs_offset(&self) -> Option<u32>

Gets the class definition list offset.

fn get_data_size(&self) -> u32

Gets the data section size.

fn get_data_offset(&self) -> u32

Gets the data secrion offset.