Struct dalvik::read::DexReader

pub struct DexReader {
    file_cursor: Cursor<Box<[u8]>>,
    header: Header,
    strings: Vec<String>,
    types: Vec<Type>,
    prototypes: Vec<Prototype>,
    field_ids: Vec<FieldIdData>,
    method_ids: Vec<MethodIdData>,
    classes: Vec<Class>,
    annotation_set_ref_list: Vec<Box<[u32]>>,
    annotation_sets: Vec<Box<[u32]>>,
    code_segments: Vec<(u32, CodeItem)>,
    debug_info: Vec<(u32, DebugInfo)>,
    arrays: Vec<(u32, Array)>,
    annotations_directories: Vec<(u32, AnnotationsDirectory)>,
}

Structure for reading a Dex file in a fast way.

Fields
file_cursor: Cursor<Box<[u8]>>

Reader to use to read data.

Header must already be read before creating the object.

header: Header

Header of the dex file.

strings: Vec<String>

String list.

types: Vec<Type>

Type list.

prototypes: Vec<Prototype>

Prototype ID list.

field_ids: Vec<FieldIdData>

Field ID list.

This list creates 1:1 relations between field indexes and field information offsets in the dex file.

method_ids: Vec<MethodIdData>

Method ID list.

classes: Vec<Class>

List of classes.

annotation_set_ref_list: Vec<Box<[u32]>>

List of lists of references to annotation set offsets.

annotation_sets: Vec<Box<[u32]>>

Set of annotations.

code_segments: Vec<(u32, CodeItem)>

Code segment list.

debug_info: Vec<(u32, DebugInfo)>

Debug information list.

arrays: Vec<(u32, Array)>

Array list.

annotations_directories: Vec<(u32, AnnotationsDirectory)>

Annotations directories.

Methods

impl DexReader

fn new<R: Read>(file: R, size: Option<usize>) -> Result<DexReader>

Creates a new reader with the information from the header of the file.

fn read_data(&mut self) -> Result<()>

Reads data from a whole file and stores its information.

fn read_endian_data<B: ByteOrder>(&mut self) -> Result<()>

Reads the data in the correct endianness.

fn read_string_list<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of strings.

fn read_string(&mut self) -> Result<String>

Reads an actual string.

fn read_all_types<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of types.

fn read_prototype_list<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of prototype IDs.

fn read_type_list<B: ByteOrder>(&mut self) -> Result<Box<[Type]>>

Reads a list of types.

fn read_field_id_list<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of field IDs.

fn read_method_id_list<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of method IDs.

fn read_class_list<B: ByteOrder>(&mut self) -> Result<()>

Reads the list of classes.

fn read_annotations_directory<B: ByteOrder>(&mut self)
-> Result<AnnotationsDirectory>

Reads an annotations directory.

fn read_annotation_set<B: ByteOrder>(&mut self) -> Result<Box<[Annotation]>>

Reads an annotation set.

fn read_annotation(&mut self) -> Result<Annotation>

Reads an annotation.