Struct dalvik::Dex

pub struct Dex {
    header: Header,
    strings: Vec<String>,
    types: Vec<String>,
}

Dex file representation.

Fields
header: Header strings: Vec<String> types: Vec<String>
Methods

impl Dex

fn from_file<P: AsRef<Path>>(path: P) -> Result<Dex>

Reads the Dex data structure from the given path.

fn from_reader<R: BufRead>(reader: R, size: Option<usize>) -> Result<Dex>

Loads a new Dex data structure from the given reader.