Struct dalvik::types::read::ClassDefData

pub struct ClassDefData {
    class_index: u32,
    access_flags: AccessFlags,
    superclass_index: Option<u32>,
    interfaces_offset: Option<u32>,
    source_file_index: Option<u32>,
    annotations_offset: Option<u32>,
    class_data_offset: Option<u32>,
    static_values_offset: Option<u32>,
}

Data of a class definition.

Fields
class_index: u32 access_flags: AccessFlags superclass_index: Option<u32> interfaces_offset: Option<u32> source_file_index: Option<u32> annotations_offset: Option<u32> class_data_offset: Option<u32> static_values_offset: Option<u32>
Methods

impl ClassDefData

fn from_reader<R: Read, B: ByteOrder>(reader: &mut R) -> Result<ClassDefData>

Creates a new ClassDefData from a reader.

fn class_index(&self) -> u32

Gets the index in the Type IDs list of the class.

fn access_flags(&self) -> AccessFlags

Gets the access flags of the class definition.

fn superclass_index(&self) -> Option<u32>

Gets the index in the Type IDs list of the superclass of this class, if it exists.

fn interfaces_offset(&self) -> Option<u32>

Gets the offset of the list of interfaces implemented by the class, if it has any.

fn source_file_index(&self) -> Option<u32>

Gets the index in the String IDs list of the file name where most of the class was, if it exists.

fn annotations_offset(&self) -> Option<u32>

Gets the offset of the annotations of the class, if it has any.

fn class_data_offset(&self) -> Option<u32>

Gets the offset for the data of the class, if it has any.

fn static_values_offset(&self) -> Option<u32>

Gets the offset for the static values of the class, if it has any.