Struct dalvik::types::PrototypeIdData

pub struct PrototypeIdData { /* fields omitted */ }

Data structure representing the proto_id_item type.

Methods

impl PrototypeIdData

fn from_reader<R: Read, E: ByteOrder>(reader: &mut R) -> Result<PrototypeIdData>

Creates a new PrototypeIdData from a reader.

fn get_shorty_index(&self) -> usize

Gets the shorty index.

Gets the index into the string_ids list for the short-form descriptor string of this prototype. The string must conform to the syntax for ShortyDescriptor, and must correspond to the return type and parameters of this item.

fn get_return_type_index(&self) -> usize

Gets the return type index.

Gets the index into the type_ids list for the return type of this prototype.

fn get_parameters_offset(&self) -> Option<u32>

Gets the parameter list offset, if the prototype has parameters.

Gets the offset from the start of the file to the list of parameter types for this prototype, or None if this prototype has no parameters. This offset, should be in the data section, and the data there should be in the format specified by type_list. Additionally, there should be no reference to the type void in the list.