Struct dalvik::header::Header

pub struct Header {
    magic: [u8; 8],
    checksum: u32,
    signature: [u8; 20],
    file_size: u32,
    header_size: u32,
    endian_tag: u32,
    link_size: u32,
    link_offset: Option<u32>,
    map_offset: u32,
    string_ids_size: u32,
    string_ids_offset: Option<u32>,
    type_ids_size: u32,
    type_ids_offset: Option<u32>,
    prototype_ids_size: u32,
    prototype_ids_offset: Option<u32>,
    field_ids_size: u32,
    field_ids_offset: Option<u32>,
    method_ids_size: u32,
    method_ids_offset: Option<u32>,
    class_defs_size: u32,
    class_defs_offset: Option<u32>,
    data_size: u32,
    data_offset: u32,
}

Dex header representation structure.

Fields
magic: [u8; 8]checksum: u32signature: [u8; 20]file_size: u32header_size: u32endian_tag: u32link_size: u32link_offset: Option<u32>map_offset: u32string_ids_size: u32string_ids_offset: Option<u32>type_ids_size: u32type_ids_offset: Option<u32>prototype_ids_size: u32prototype_ids_offset: Option<u32>field_ids_size: u32field_ids_offset: Option<u32>method_ids_size: u32method_ids_offset: Option<u32>class_defs_size: u32class_defs_offset: Option<u32>data_size: u32data_offset: u32
Methods

impl Header

pub fn from_file<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Obtains the header from a Dex file.

pub fn from_reader<R: Read>(reader: R) -> Result<Self, Error>

Obtains the header from a Dex file reader.

fn read_data<R: Read, E: ByteOrder>(
reader: R,
magic: [u8; 8],
checksum: u32,
signature: [u8; 20],
file_size: u32,
header_size: u32,
endian_tag: u32
) -> Result<Self, Error>

fn is_magic_valid(magic: [u8; 8]) -> bool

Checks if the dex magic number given is valid.

pub fn get_magic(&self) -> &[u8; 8]

Gets the magic value.

pub fn get_dex_version(&self) -> u8

Gets Dex version.

pub fn get_checksum(&self) -> u32

Gets file checksum.

pub fn get_signature(&self) -> &[u8; 20]

Gets file SHA-1 signature.

pub fn get_file_size(&self) -> u32

Gets file size.

pub fn get_header_size(&self) -> u32

Gets header size, in bytes.

This must be 0x70.

pub fn get_endian_tag(&self) -> u32

Gets the endian tag.

This must be ENDIAN_CONSTANT or REVERSE_ENDIAN_CONSTANT.

pub fn is_little_endian(&self) -> bool

Gets wether the file is in little endian or not.

pub fn is_big_endian(&self) -> bool

Gets wether the file is in big endian or not.

pub fn get_link_size(&self) -> u32

Gets the link section size

pub fn get_link_offset(&self) -> Option<u32>

Gets the link section offset.

pub fn get_map_offset(&self) -> u32

Gets the map section offset.

pub fn get_string_ids_size(&self) -> u32

Gets the string IDs list size.

pub fn get_string_ids_offset(&self) -> Option<u32>

Gets the string IDs list offset.

pub fn get_type_ids_size(&self) -> u32

Gets the type IDs list size.

pub fn get_type_ids_offset(&self) -> Option<u32>

Gets the type IDs list offset.

pub fn get_prototype_ids_size(&self) -> u32

Gets the prototype IDs list size.

pub fn get_prototype_ids_offset(&self) -> Option<u32>

Gets the prototype IDs list offset.

pub fn get_field_ids_size(&self) -> u32

Gets the field IDs list size.

pub fn get_field_ids_offset(&self) -> Option<u32>

Gets the field IDs list offset.

pub fn get_method_ids_size(&self) -> u32

Gets the method IDs list size.

pub fn get_method_ids_offset(&self) -> Option<u32>

Gets the method IDs list offset.

pub fn get_class_defs_size(&self) -> u32

Gets the class definition list size.

pub fn get_class_defs_offset(&self) -> Option<u32>

Gets the class definition list offset.

pub fn get_data_size(&self) -> u32

Gets the data section size.

pub fn get_data_offset(&self) -> u32

Gets the data section offset.