Struct dalvik::read::DexReader

pub struct DexReader<R: BufRead> {
    reader: R,
    current_offset: u32,
    header: Header,
    offset_map: OffsetMap,
    unknown_data: Vec<(u32, Box<[u8]>)>,
    string_ids: Vec<u32>,
    type_ids: Vec<u32>,
    prototype_ids: Vec<PrototypeIdData>,
    field_ids: Vec<FieldIdData>,
    method_ids: Vec<MethodIdData>,
    class_defs: Vec<ClassDefData>,
    type_lists: Vec<Box<[u16]>>,
    annotation_set_ref_list: Vec<Box<[u32]>>,
    annotation_sets: Vec<Box<[u32]>>,
    classes: Vec<(u32, ClassData)>,
    code_segments: Vec<(u32, CodeItem)>,
    strings: Vec<(u32, String)>,
    debug_info: Vec<(u32, DebugInfo)>,
    annotations: Vec<(u32, AnnotationItem)>,
    arrays: Vec<(u32, Array)>,
    annotations_directories: Vec<(u32, AnnotationsDirectory)>,
}

Structure for reading a Dex file in a fast way.

Fields
reader: R

Reader to use to read data.

Header must already be read before creating the object.

current_offset: u32

Current reading offset.

header: Header

Header of the dex file.

offset_map: OffsetMap

Map to store all known offsets in the Dex file.

unknown_data: Vec<(u32, Box<[u8]>)>

Vector of chunks of binary data that were not understood when reading sequencially.

string_ids: Vec<u32>

String ID list.

This list creates 1:1 relations between String indexes and string offsets in the dex file.

type_ids: Vec<u32>

Type ID list.

This list creates 1:1 relations between type indexes and type descriptor string index.

prototype_ids: Vec<PrototypeIdData>

Prototype ID list.

field_ids: Vec<FieldIdData>

Field ID list.

This list creates 1:1 relations between field indexes and field information offsets in the dex file.

method_ids: Vec<MethodIdData>

Method ID list.

class_defs: Vec<ClassDefData>

Class definition list

type_lists: Vec<Box<[u16]>>

Type list.

Each element is the index in the type ID list for the type.

annotation_set_ref_list: Vec<Box<[u32]>>

List of lists of references to annotation set offsets.

annotation_sets: Vec<Box<[u32]>>

Set of annotations.

classes: Vec<(u32, ClassData)>

Class data list.

code_segments: Vec<(u32, CodeItem)>

Code segment list.

strings: Vec<(u32, String)>

String list.

debug_info: Vec<(u32, DebugInfo)>

Debug information list.

annotations: Vec<(u32, AnnotationItem)>

Annotation list.

arrays: Vec<(u32, Array)>

Array list.

annotations_directories: Vec<(u32, AnnotationsDirectory)>

Annotations directories.

Methods

impl<R> DexReader<R> where R: BufRead

fn new(header: Header, reader: R) -> DexReader<R>

Creates a new reader with the information from the header of the file.

fn read_data<E: ByteOrder>(&mut self) -> Result<()>

Reades data from a whole file and stores its information.

fn read_next<E: ByteOrder>(&mut self, offset_type: OffsetType) -> Result<()>

Reads the next element

fn read_string_id_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of string IDs.

fn read_type_id_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of type IDs.

fn read_prototype_id_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of prototype IDs.

fn read_field_id_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of field IDs.

fn read_method_id_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of method IDs.

fn read_class_def_list<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of class definitions.

fn read_map<E: ByteOrder>(&mut self) -> Result<()>

Reads the map of the dex file.

fn read_type_lists<E: ByteOrder>(&mut self) -> Result<()>

Reads the list of types.

fn read_annotation_set_list<E: ByteOrder>(&mut self) -> Result<()>

Reads a list of annotation sets.

fn read_annotation_set<E: ByteOrder>(&mut self) -> Result<()>

Reads an annotation set.

fn read_class_data(&mut self) -> Result<()>

Reads the class information.

fn read_code_item<E: ByteOrder>(&mut self) -> Result<()>

Reads code information.

fn read_string_data(&mut self) -> Result<()>

Reads an actual string.

fn read_debug_info(&mut self) -> Result<()>

Reads debug information.

fn read_annotation(&mut self) -> Result<()>

Reads an annotation.

fn read_encoded_array(&mut self) -> Result<()>

Reads an encoded array.

fn read_annotation_directory<E: ByteOrder>(&mut self) -> Result<()>