Struct dalvik::Dex

pub struct Dex {
    header: Header,
    strings: Vec<String>,
    types: Vec<String>,
    prototypes: Vec<Prototype>,
    fields: Vec<Field>,
    methods: Vec<Method>,
    classes: Vec<ClassDef>,
}

Fields
header: Header strings: Vec<String> types: Vec<String> prototypes: Vec<Prototype> fields: Vec<Field> methods: Vec<Method> classes: Vec<ClassDef>
Methods

impl Dex

fn from_file<P: AsRef<Path>>(path: P) -> Result<Dex>

Reads the Dex data structure from the given path.

fn from_reader<R: BufRead>(reader: R, file_size: u64) -> Result<Dex>

Loads a new Dex data structure from the given reader.

fn verify_file<P: AsRef<Path>>(&self, path: P) -> bool

Verifies the file at the given path.

fn verify_reader<R: Read>(&self, reader: R) -> bool

Verifies the file in the given reader.

The reader should be positioned at the start of the file.