Struct dalvik::types::Class

pub struct Class {
    class_index: u32,
    access_flags: AccessFlags,
    superclass_index: Option<u32>,
    interfaces: Box<[Type]>,
    source_file_index: Option<u32>,
    annotations: Option<AnnotationsDirectory>,
    class_data: Option<ClassData>,
    static_values: Option<Array>,
}

Structure representing a class.

Fields
class_index: u32access_flags: AccessFlagssuperclass_index: Option<u32>interfaces: Box<[Type]>source_file_index: Option<u32>annotations: Option<AnnotationsDirectory>class_data: Option<ClassData>static_values: Option<Array>
Methods

impl Class

pub fn new(
class_index: u32,
access_flags: AccessFlags,
superclass_index: Option<u32>,
interfaces: Box<[Type]>,
source_file_index: Option<u32>,
annotations: Option<AnnotationsDirectory>,
class_data: Option<ClassData>,
static_values: Option<Array>
) -> Self

Creates a new class.

pub fn class_index(&self) -> u32

Gets the index of the class in the type IDs list.

pub fn access_flags(&self) -> AccessFlags

Gets the access flags of the class.

pub fn superclass_index(&self) -> Option<u32>

Gets the index in the type IDs list of the superclass for this class, ifd it exists.

pub fn interfaces(&self) -> &[Type]

Gets the list of interfaces implemented by the class.

pub fn source_file_index(&self) -> Option<u32>

Gets the index of the source file in the string list if it's known.

pub fn annotations(&self) -> Option<&AnnotationsDirectory>

Gets the annotations for the class, if there are any.

pub fn class_data(&self) -> Option<&ClassData>

Gets the data associated with the class.

pub fn static_values(&self) -> Option<&Array>

Gets the arrays with the values for the static files in this class.

The values are in the same order as the static_field_ids in the class data of the class. If a value is not found, it is considered 0 or NULL depending on the type of the variable.